Anonymous Function and Closure Difference
An anonymous function is just a function that has no name; nothing more. A closure is a function that captures the state of the surrounding environment.

An anonymous function does not necessarily need to create a closure, and a closure is not created only for anonymous functions.

Consider this hypothetical counter-example. Consider a language Foo which does not support closures but supports anonymous functions. This language may either not compile or throw an error for the code below because "greeting" is not defined in the scope of the inner function. The fact that it is anonymous is irrelevant.

The above code is all about how closures are different and useful.

Although in the code the inner function is not anonymous anymore, it still captures state from the surrounding environment.
Closures provide much needed convenience, as otherwise we would be passing every single dependency of the function as an argument.